backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlmodel==0.0.22
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
httpx==0.28.1
google-generativeai==0.8.4
